% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annualtemps}
\alias{annualtemps}
\title{2010-2099 global projected annual average temperature anomalies}
\format{A data table with 55080 rows and 4 variables:
\describe{
  \item{lon}{longitude}
  \item{lat}{latitude}
  \item{Year}{year}
  \item{z}{annual average temperature in degrees Celcius}
}}
\source{
\url{https://github.com/leonawicz/mapmate}
}
\usage{
annualtemps
}
\description{
A data set containing low spatial resolution 2010-2099 global projected annual average temperatures.
This is a toy data set used for example purposes.
Anomalies are element-wise (cell by cell) delta change from a historical baseline.
}
\keyword{datasets}
