% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bathymetry}
\alias{bathymetry}
\title{Spatially aggregated example global bathymetry surface}
\format{A data table with 5832 rows and 3 variables:
\describe{
  \item{long}{longitude}
  \item{lat}{latitude}
  \item{z}{depth below sea level}
}}
\source{
\url{https://cran.r-project.org/web/packages/marmap/index.html}
}
\usage{
bathymetry
}
\description{
A bathymetry dataset containing columns of longitude, latitude and depths/elevations relative to sea level,
modified from the 10-minute resolution source bathymetry map downloaded via the \code{marmap} package.
}
\keyword{datasets}

