% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{borders}
\alias{borders}
\title{Global national political borders}
\format{A data frame with 99338 rows and 6 variables:
\describe{
  \item{lon}{longitude}
  \item{lat}{latitude}
  \item{group}{grouping for polygon lines}
  \item{order}{order of lines in group}
  \item{region}{boundary regions}
  \item{subregion}{boundary subregions}
}}
\source{
\url{https://cran.r-project.org/web/packages/maps/index.html}
}
\usage{
borders
}
\description{
A dataset containing global national political boundary lines, modified from the world map in the \code{maps} package.
}
\keyword{datasets}

