% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{do_projection}
\alias{do_projection}
\title{Project points onto globe}
\usage{
do_projection(x, lon = 0, lat = 0, n.period = 360, n.frames = n.period)
}
\arguments{
\item{x}{a data frame.}

\item{lon}{starting longitude for rotation sequence or vector of arbitrary longitude sequence.}

\item{lat}{fixed latitude or vector of arbitrary latitude sequence.}

\item{n.period}{intended length of the period.}

\item{n.frames}{intended number of frames in animation.}
}
\value{
returns a data frame containing visible points on the globe.
}
\description{
Project points in \code{x} onto the globe and filter \code{x} to points within the current field of view.
}
\details{
\code{do_projection} projects the coordinates in \code{x} onto the globe and filters \code{x} to the subset of rows
containing data which are visible given the current field of view.
The field of view is defined by the centroid focus latitude and longitude pair in the sequence of latitudes and longitudes whose index
corresponds to the frame ID in \code{x}.
}
\examples{
# not run
}

