% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_clim}
\alias{get_clim}
\title{Obtain a list of monthly climatologies}
\usage{
get_clim(x, limits = c(1961, 1990))
}
\arguments{
\item{x}{a data frame.}

\item{limits}{inclusive lower and upper bound limits for subsetting the years (climatology period).}
}
\value{
returns a list of data frames.
}
\description{
\code{get_clim} subsets and summarizes a data frame of monthly map data over a specified time period.
}
\details{
\code{get_clim} takes a five-column data frames containing map data (long, lat, z, Year, Month) describing a temporal sequence of map data.
The data frame is subset to a specified range of years. The data (z) is summarized over time for each month and a length-12 list of data frames containing period average map data is returned.
}
\examples{
#not run
}

