% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_lonlat_seq}
\alias{get_lonlat_seq}
\title{Generate a sequence of coordinates}
\usage{
get_lonlat_seq(lon, lat, n.period = 360, n.frames = n.period)
}
\arguments{
\item{lon}{vector of arbitrary longitudes or starting longitude (starting point for constant globe rotation).}

\item{lat}{vector of arbitrary latitudes or fixed, repeating latitude.}

\item{n.period}{intended length of the period.}

\item{n.frames}{intended number of frames in animation.}
}
\value{
returns a list containing a longitude vector and a latitude vector.
}
\description{
Generate a repeating sequence of longitude and latitude coordinates based on a period of rotation.
}
\details{
\code{lon} and \code{lat} may be scalars or vectors. If scalar, \code{lat} is simply repeated \code{n.frames} times.
If scalar, \code{lon} is always treated as a starting longitude and a rotational sequence of longitudes is beginning from \code{lon}
is generated for a length equal to \code{n.frames}, repeating the sequence if necessary.
If \code{lon} or \code{lat} are vectors, it is assumed that predefined custom sequences of longitude and latitude have been provided,
so they are required to be of length \code{n.period}.
However long it takes to iterate through the custom coordinates sequence ought to define the period.
It also makes sense in this case for \code{n.frames} to remain equal to \code{n.period} if the custom sequence is not meant to be cyclical but rather a single pass.
After this check the custom vectors are simply bound in a list.
}
\examples{
# not run
}

