% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_ma}
\alias{get_ma}
\title{Obtain moving average map series}
\usage{
get_ma(x, type, season = NULL, size = 10, n.cores = 32)
}
\arguments{
\item{x}{list of data frames.}

\item{type}{character, one of \code{"monthly"}, \code{"annual"}, or \code{"seasonal"}.}

\item{season}{\code{NULL} or character, one of \code{"winter"}, \code{"spring"}, \code{"summer"}, or \code{"autumn"}. Default is \code{NULL}.}

\item{size}{number of years for the moving average window. Default \code{10}.}

\item{n.cores}{number of CPUs for parallel processing.}
}
\value{
returns a list of data frames.
}
\description{
Obtain a moving average for monthly, annual or seasonal resolution data from monthly map data.
}
\details{
\code{get_ma} takes a list of data frames and for each data frame computes and returns the moving or rolling average
after first summarizing monthly data to seasonal or annual averages if applicable.
Winter begins in December and each season is three consecutive months (DJF, MAM, JJA, SON).
Parallel processing is Linux-only (uses \code{mclapply}) and the default of 32 CPU cores will have to be changed if you don't have that many.
It is convenient for my usage and this package version is not intended for general use.
Most other users will probably not be using parallel processing at all, in which case \code{n.cores} is ignored.
}
\examples{
# not run
}

