% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monthlytemps}
\alias{monthlytemps}
\title{2010-2099 single map grid cell projected monthly average temperatures anomalies}
\format{A data table with 1080 rows and 5 variables:
\describe{
  \item{lon}{longitude}
  \item{lat}{latitude}
  \item{Year}{year}
  \item{Month}{January-December month index 01 through 12}
  \item{z}{annual average temperature in degrees Celcius}
}}
\source{
\url{https://github.com/leonawicz/mapmate}
}
\usage{
monthlytemps
}
\description{
A data set containing 2010-2099 projected annual average temperature anomalies for a single low resolution grid cell
taken from a global monthly data set also used as the basis for the \code{annualtemps} data set in this package.
This is a toy data set used for example purposes.
This data set includes data from only one grid cell because it is sufficient for examples but also keeps the size of the data small.
Anomalies are delta change from a historical baseline.
}
\keyword{datasets}
