% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network}
\alias{network}
\title{Simulated set of world cities locations and population-based weights}
\format{A data table with 1000 rows and 3 variables:
\describe{
  \item{lon}{longitude}
  \item{lat}{latitude}
  \item{Pop_wts}{population-based weights}
}}
\source{
\url{https://github.com/leonawicz/mapmate}
}
\usage{
network
}
\description{
A data set containing simulated world cities locations and population-based weights.
This is a toy data set used for example purposes.
Sampling and weights are based on the square root of the original populations and are rescaled and rounded.
}
\keyword{datasets}

