% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runOrtho.R
\name{runOrtho}
\alias{runOrtho}
\title{Run Shiny app example}
\usage{
runOrtho()
}
\description{
Launch a Shiny app which provides an example of the effect of longitude, latitude and axis orientation argument settings on the displayed orthographic globe projection.
}
\examples{
# not run
}

