% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{theme_blank_plus}
\alias{theme_blank_plus}
\title{Blank ggplot2 theme with optional axes}
\usage{
theme_blank_plus(col = "transparent")
}
\arguments{
\item{col}{axis line, tick and text color. Defaults to \code{"transparent"}.}
}
\description{
A blank ggplot2 theme which will draw only data, but can include axes lines, ticks, and text is the color is not set to transparent.
}
\details{
\code{theme_blank_plus} is intended for plotting data, e.g., a line plot on a blank canvas (nothing drawn but the line itself)
while still retaining space for axes which may be added later so that everything will line up easily in overlaid plots.
Hence, the other time it is used is to plot visible axes, but perhaps no data.

A common use case is as follows: use \code{theme_blank_plus} repeatedly while saving high-resolution images to disk of a sequence of time series plots
where the time series line grows from left to right through the still image sequence. The axes are fixed across all plots so there is no need to draw them every time, but space is left for them.
Subsequently, only a simgle plot is saved to disk of the axes (with no data drawn) and this image in layered with the sequence of data images when an animation is made.
}
\examples{
# not run
}

