% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{project_to_hemisphere}
\alias{project_to_hemisphere}
\title{Identidy visible points on an arbitrary global hemishpere view.}
\usage{
project_to_hemisphere(lat, long, lat0, long0)
}
\arguments{
\item{lat}{vector of latitudes.}

\item{long}{vector of longitudes.}

\item{lat0}{latitude of focus coordinates.}

\item{long0}{longitude of focus coordinates.}
}
\value{
returns a data table.
}
\description{
Given a global hemispheric field of view defined by a single latitudinal and longitudinal centroid focal point, project geographic points onto the hemishpere.
}
\details{
\code{project_to_hemisphere} identifies whether each pair of coordinates in the \code{lat} and \code{long} vectors is in a field of view defined by a centroid focal point \code{(lat0, long0)}
and returns a data table containing the original coordinates and a column indicating if the coordinates are in the field of view (\code{TRUE} or \code{FALSE}).
}
\examples{
#not run
}

