% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{save_map}
\alias{save_map}
\title{Save maps to disk}
\usage{
save_map(x, dir = getwd(), lon = 0, lat = 0, n.period = 360,
  n.frames = n.period, ortho = TRUE, col = NULL, type = "network",
  suffix = NULL, z.range = NULL, rotation.axis = 23.4,
  png.args = list(width = 1920, height = 1080, res = 300, bg = "transparent"),
  save.plot = TRUE, return.plot = FALSE, num.format = 4)
}
\arguments{
\item{x}{a data frame containing networks, tiles, or lines information.}

\item{dir}{png output directory. Defaults to working directory.}

\item{lon}{starting longitude for rotation sequence or vector of arbitrary longitude sequence.}

\item{lat}{fixed latitude or vector of arbitrary latitude sequence.}

\item{n.period}{intended length of the period.}

\item{n.frames}{intended number of frames in animation.}

\item{ortho}{use an orthographic projection for globe plots. Defaults to \code{TRUE}.}

\item{col}{sensible default colors provided for each \code{type}}

\item{type}{the type of plot, one of \code{"network"}, \code{maptiles}, or \code{maplines}.}

\item{suffix}{character, optional suffix to be pasted onto output filename.}

\item{z.range}{the full known range for the data values across all \code{x} objects, not just the current one.}

\item{rotation.axis}{the rotation axis used when \code{ortho=TRUE} for globe plots. Defaults to 23.4 degrees.}

\item{png.args}{a list of arguments passed to \code{png}.}

\item{save.plot}{save the plot to disk. Defaults to \code{TRUE}. Typically only set to \code{FALSE} for demonstrations and testing.}

\item{return.plot}{return the ggplot object. Defaults to \code{FALSE}. Only intended for single-plot demonstrations and testing, not for still image sequence automation.}

\item{num.format}{number of digits including any leading zeros for image sequence frame numbering. Defaults to 4, i.e. \code{0001, 0002, ...}.}
}
\value{
usually returns NULL after writing file to disk as a side effect. May return a ggplot object but be careful not to use this option if looping over many plots.
}
\description{
Save a map to disk intended to be part of a as a still image sequence of one of three types: networks, tiles, or lines.
}
\details{
\code{save_map} takes a specific type of data frame catering to networks, tiles, or lines.
It plots a 3D globe map with \code{ortho=TRUE} (default) or a flat map (\code{ortho=FALSE}).
For flat maps, \code{lon}, \code{lat}, \code{n.period}, \code{n.frames}, and \code{rotation.axis} are ignored.
For plotting on a globe, \code{lon} and \code{lat} are used to describe the field of view or the visible hemisphere.
\code{n.period} relates is eithe the period of rotation of the globe or the length of the non-repeating, arbitrary coordinates sequence.
\code{n.frames} is always the explicit number of frames that will make up an animation
regardless of the length of the series of data frames \code{x} to be plotted or the length of the rotational period or coordinates sequence.

\code{z.range} is important for \code{type="maptiles"} because it is used to ensure colors are mapped to values consistently across all plots.
This is not only for the case of changing data values across a series of plots of different data frames \code{x}.
There are also changes in the range of values for a fixed data frame when it is plotted repeatedly as the globe is rotated and different hemispheres of the map
(different data subsets) are in view across the image sequence.

The png output directory will be created if it does not exist, recursively if necessary. The default is the working directory.
This is ignored if \code{save.plot=FALSE}.
}
\examples{
# not run
}

