% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{save_ts}
\alias{save_ts}
\title{Save time series plots}
\usage{
save_ts(i, x, label, col, xlm, ylm, axes_only = FALSE)
}
\arguments{
\item{i}{current time index used to subset \code{x}.}

\item{x}{data frame for plotting, currently assuming columns of \code{Year} and \code{Mean} as the x and y plotting variables.}

\item{label}{a subdirectory label. Plots are saved in the working directory under \code{frames/{label}}.}

\item{col}{color of the time series line or the axes lines, ticks, and text.}

\item{xlm}{x axis limits.}

\item{ylm}{y axis limits.}

\item{axes_only}{only plot axis information, no data. Defaults to \code{FALSE}.}
}
\description{
Save a time series plot to disk intended to be part of a as a still image sequence of a growing time series.
}
\examples{
# not run
}

